fixed project file structure
