+++ client/src/App.js
@@ -1,48 +1,29 @@
 import React from 'react';
-import styled from 'styled-components';
 import { Route, Switch } from 'react-router-dom';
-import ListOfItems from './Pages/ListOfItems';
-import Cart from './Pages/Cart';
-import Filter from './Pages/Filter';
-import Profile from './Pages/Profile';
-import Header from './Components/Header';
-import Multicarousel from './Components/Carousels/Multicarousel';
+import CartPage from './Pages/CartPage';
+import CatalogPage from './Pages/CatalogPage';
+import ProfilePage from './Pages/ProfilePage';
+import IndexPage from './Pages/IndexPage';
 
 function App() {
 	return (
 		<div>
-			<AppContainer>
-				<Header />
-				<MulticarouselContainer>
-					<Multicarousel />
-				</MulticarouselContainer>
-				<Header />
-
-				<Switch>
-					<Route exact path="/">
-						<ListOfItems />
-					</Route>
-					<Route path="/cart">
-						<Cart />
-					</Route>
-					<Route path="/filter">
-						<Filter />
-					</Route>
-					<Route path="/profile">
-						<Profile />
-					</Route>
-				</Switch>
-			</AppContainer>
+			<Switch>
+				<Route exact path="/">
+					<IndexPage />
+				</Route>
+				<Route path="/cart">
+					<CartPage />
+				</Route>
+				<Route path="/filter">
+					<CatalogPage />
+				</Route>
+				<Route path="/profile">
+					<ProfilePage />
+				</Route>
+			</Switch>
 		</div>
 	);
 }
-
-const AppContainer = styled.div`
-	width: 90%;
-	margin: auto;
-`;
-const MulticarouselContainer = styled.div`
-	margin: 20px 0 20px 0;
-`;
 
 export default App;
