+++ client/src/Components/Carousels/ProductsCarousel/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const ProductCardContainer = styled.div`
+width: 80%;
+margin: auto;
+`;
