+++ client/src/Components/Header/styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Nav = styled.nav`
+	display: flex;
+	justify-content: space-around;
+	width: 100%;
+`;
