+++ client/src/Components/UI/InputFiels/styles.js
@@ -1,0 +1,22 @@
+import styled from 'styled-components';
+
+export const InputContainer = styled.div`
+	font-family: 'Raleway', sans-serif;
+	margin-bottom: 22px;
+`;
+
+export const Input = styled.input`
+	display: inline-block;
+	border: 1px solid #16161c;
+	box-sizing: border-box;
+	border-radius: 4px;
+	padding: 6px 12px;
+	font-family: Raleway;
+	font-size: 16px;
+	outline: none;
+`;
+
+export const Label = styled.label`
+	display: block;
+	margin-bottom: 5px;
+`;
