--- client/src/Pages/Cart.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-
-
-export default function Cart() {
-    return(
-        <>
-        it is the cart
-        </>
-    )
-}
