+++ client/src/Pages/CartPage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CartPage = () => {
+	return <div>It is CartPage</div>;
+};
+
+export default CartPage;
