+++ client/src/Pages/CatalogPage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CatalogPage = () => {
+	return <div>It is CatalogPage!</div>;
+};
+
+export default CatalogPage;
