--- client/src/Pages/Filter.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Filter() {
-    return(
-        <>
-        it is the filter
-        </>
-    )
-}
