+++ client/src/Pages/IndexPage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../Components/Header/Header';
+import Multicarousel from '../Components/Carousels/ProductsCarousel/ProductsCarousel';
+
+const IndexPage = () => {
+	return (
+		<>
+			<Header />
+			<Multicarousel />
+		</>
+	);
+};
+
+export default IndexPage;
