--- client/src/Pages/ListOfItems.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function ListOfItems() {
-    return(
-        <>
-        it is the list of items
-        </>
-    )
-}
