--- client/src/Pages/Profile.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Profile() {
-    return(
-        <>
-        it is the profile
-        </>
-    )
-}
