+++ client/src/Pages/ProfilePage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ProfilePage = () => {
+	return <>It is ProfilePage</>;
+};
+
+export default ProfilePage;
